requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
python-multipart 
streamlit
pywin32 

pandas
ragas

langchain
langchain-community
langchain-huggingface 
langchain-google-genai 
langchain-groq 
langchain-pinecone
langchain-chroma
langchain-qdrant
qdrant-client
hnswlib